Prevent internal storage from filling up
Ridiculous configurations can no longer cause vending machine internal storage to fill up completely. This is ensured by keeping on only one stack of each item.

This comes with the side effect of allowing items to be purchased beyond the server's configured stack size, which will be addressed in a later change.

The VendingInStock plugin can bypass this mitigation, which will be addressed in a later change.