Fix VendingInStock compatibility and support genetics
This plugin now respects the VendingInStock configuration, rather than simply instantly restocking all items while VendingInStock is loaded.

Saving a plant will capture its genetics information and ensure players are given a plant item with the same genetics. Players won't be able to see the genetics information in the shop UI, but that can be addressed in a future update by overlaying the genetics information like how vanilla does it.